+++ src/SFA.DAS.Campaigns.UITests/Helpers/CampaignsStepsHelper.cs
@@ -1,4 +1,5 @@
 ﻿using SFA.DAS.Campaigns.UITests.Project.Tests.Pages;
+using SFA.DAS.Campaigns.UITests.Project.Tests.Pages.Apprentices;
 using System.Threading.Tasks;
 using TechTalk.SpecFlow;
