+++ src/SFA.DAS.Campaigns.UITests/Project/GlobalUsing.cs
@@ -1,0 +1,10 @@
+﻿global using Microsoft.Playwright;
+global using NUnit.Framework;
+global using SFA.DAS.Campaigns.UITests.Helpers;
+global using SFA.DAS.ConfigurationBuilder;
+global using SFA.DAS.Framework;
+global using SFA.DAS.FrameworkHelpers;
+global using System;
+global using System.Collections.Generic;
+global using System.Threading.Tasks;
+global using TechTalk.SpecFlow;
