--- src/SFA.DAS.Campaigns.UITests/Project/Tests/Pages/ApprenticeHomePage.cs
@@ -1,11 +1,0 @@
-﻿using Microsoft.Playwright;
-using System.Threading.Tasks;
-using TechTalk.SpecFlow;
-
-namespace SFA.DAS.Campaigns.UITests.Project.Tests.Pages
-{
-    public class ApprenticeHomePage(ScenarioContext context) : CampaingnsBasePage(context)
-    {
-        public override async Task VerifyPage() => await Assertions.Expect(driver.Page.Locator("h1")).ToContainTextAsync("Become an apprentice");
-    }
-}
