+++ src/SFA.DAS.Campaigns.UITests/Project/Tests/Pages/CampaingnsBasePage.cs
@@ -1,12 +1,6 @@
-﻿using SFA.DAS.Campaigns.UITests.Helpers;
-using SFA.DAS.Framework;
-using TechTalk.SpecFlow;
+﻿namespace SFA.DAS.Campaigns.UITests.Project.Tests.Pages;
 
-namespace SFA.DAS.Campaigns.UITests.Project.Tests.Pages
+public abstract class CampaingnsBasePage(ScenarioContext context) : BasePage(context)
 {
-
-    public abstract class CampaingnsBasePage(ScenarioContext context) : BasePage(context)
-    {
-        protected readonly CampaignsDataHelper campaignsDataHelper = context.Get<CampaignsDataHelper>();
-    }
+    protected readonly CampaignsDataHelper campaignsDataHelper = context.Get<CampaignsDataHelper>();
 }
