+++ src/SFA.DAS.Campaigns.UITests/Project/Tests/StepDefinitions/CampaignsHomeSteps.cs
@@ -1,5 +1,5 @@
 ﻿using SFA.DAS.Campaigns.UITests.Helpers;
-using SFA.DAS.Campaigns.UITests.Project.Tests.Pages;
+using SFA.DAS.Campaigns.UITests.Project.Tests.Pages.Apprentices;
 using System.Threading.Tasks;
 using TechTalk.SpecFlow;
